feedback and add another shutdown test file
